fixes lint issues
